+++ back/db/schema.rb
@@ -259,6 +259,7 @@ ActiveRecord::Schema.define(version: 2022_12_05_112729) do
     t.integer "maximum"
     t.jsonb "minimum_label_multiloc", default: {}, null: false
     t.jsonb "maximum_label_multiloc", default: {}, null: false
+    t.jsonb "logic", default: {}, null: false
     t.index ["resource_type", "resource_id"], name: "index_custom_fields_on_resource_type_and_resource_id"
   end